Explicit typecast for Curl_debug() size argument

+++ lib/http.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: http.c,v 1.300 2006-10-17 02:31:06 yangtse Exp $
+ * $Id: http.c,v 1.301 2006-10-17 10:04:14 yangtse Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -901,10 +901,11 @@ CURLcode add_buffer_send(send_buffer *in,
     if(conn->data->set.verbose) {
       /* this data _may_ contain binary stuff */
       Curl_debug(conn->data, CURLINFO_HEADER_OUT, ptr,
-                 amount-included_body_bytes, conn);
+                 (size_t)(amount-included_body_bytes), conn);
       if (included_body_bytes)
-        Curl_debug(conn->data, CURLINFO_DATA_OUT,
-                   ptr+amount-included_body_bytes, included_body_bytes, conn);
+        Curl_debug(conn->data, CURLINFO_DATA_OUT, 
+                   ptr+amount-included_body_bytes, 
+                   (size_t)included_body_bytes, conn);
     }
 
     *bytes_written += amount;
@@ -1257,8 +1258,8 @@ CURLcode Curl_proxyCONNECT(struct connectdata *conn,
 
               /* output debug if that is requested */
               if(data->set.verbose)
-                Curl_debug(data, CURLINFO_HEADER_IN, line_start, perline,
-                           conn);
+                Curl_debug(data, CURLINFO_HEADER_IN, 
+                           line_start, (size_t)perline, conn);
 
               /* send the header to the callback */
               writetype = CLIENTWRITE_HEADER;
